requirements.txt
sentence-transformers
faiss-cpu
python-dotenv
langchain
langchain-community
langchain-google-genai
numpy
tqdm
fastapi
uvicorn
langchain-huggingface
hf_xet
